+++ src/Drivers/Cab.php
@@ -40,13 +40,13 @@ class Cab extends BasicDriver
         switch ($format) {
             case Formats::CAB:
                 $abilities = [
-                    Formats::OPEN,
+                    BasicDriver::OPEN,
                 ];
 
                 $parts = explode('.', PHP_VERSION);
                 // not supported on versions below 7.0.22, 7.1.8, 7.2.0
                 if ($parts[0] > 7 || $parts[1] >= 2 || (($parts[1] == 1 && $parts[2] >= 8) || ($parts[1] == 0 && $parts[2] >= 22))) {
-                    $abilities[] = Formats::EXTRACT_CONTENT;
+                    $abilities[] = BasicDriver::EXTRACT_CONTENT;
                 }
 
                 return $abilities;