+++ src/Drivers/OneFile/Lzma.php
@@ -1,6 +1,7 @@
 <?php
 namespace wapmorgan\UnifiedArchive\Drivers\OneFile;
 
+use wapmorgan\UnifiedArchive\Drivers\BasicDriver;
 use wapmorgan\UnifiedArchive\Formats;
 use wapmorgan\UnifiedArchive\Drivers\OneFile\OneFileDriver;
 
@@ -35,7 +36,7 @@ class Lzma extends OneFileDriver
         }
         switch ($format) {
             case Formats::LZMA:
-                return [Formats::OPEN, Formats::EXTRACT_CONTENT, Formats::STREAM_CONTENT, Formats::CREATE];
+                return [BasicDriver::OPEN, BasicDriver::EXTRACT_CONTENT, BasicDriver::STREAM_CONTENT, BasicDriver::CREATE];
         }
     }